Delete TexturAtlasCompiler
