--- Runtime/TexturAtlas/TexturAtlasCompiler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 47d677cafa47a594b86cb71f4f68ecd6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
